:truck: move add event form props type
