+++ frontend/src/views/components/event/add/BasicInfomationsPanel.tsx
@@ -13,7 +13,7 @@ import ExpansionPanel from '@material-ui/core/ExpansionPanel';
 import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
 import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import Typography from '@material-ui/core/Typography';
-import { AddEventFormProps } from './AddEventFormProps';
+import { AddEventFormProps } from '../../../forms/types/AddEventFormProps';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import { makeStyles } from '@material-ui/core/styles';
