+++ frontend/src/views/components/event/add/ThrowingMoneyPanel.tsx
@@ -10,7 +10,7 @@ import ErrorMessage from '../../form/ErrorMessage';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import { AddEventFormProps } from './AddEventFormProps';
+import { AddEventFormProps } from '../../../forms/types/AddEventFormProps';
 import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles({
